Update patch set 4
Patch Set 4: Code-Review+1

This is a definite improvement from File Set 3. I'm going to keep writing what I find, I'll let you decide when it's enough of an improvement to be merged.

Let's talk about fmt.(*pp).printArg, in particular this:

 <2><311f3>: Abbrev Number: 7 (DW_TAG_formal_parameter)
    <311f4>   DW_AT_name        : p
    <311f6>   DW_AT_decl_line   : 604
    <311f8>   DW_AT_location    : 0x67cd1 (location list)
    <311fc>   DW_AT_type        : <0x40788>

Loclist is:
    00067cd1 ffffffffffffffff 00000000004827d0 (base address)
    00067ce1 00000000004827f1 000000000048290f (DW_OP_call_frame_cfa)
    00067cf4 00000000004829f7 0000000000482a35 (DW_OP_call_frame_cfa)
    00067d07 0000000000482beb 0000000000482c29 (DW_OP_call_frame_cfa)
    00067d1a 0000000000482cab 0000000000482cfa (DW_OP_call_frame_cfa)
    00067d2d 0000000000482d64 0000000000482dd3 (DW_OP_call_frame_cfa)
    00067d40 0000000000482e43 0000000000482e8b (DW_OP_call_frame_cfa)
    00067d53 0000000000482ef0 0000000000482f4d (DW_OP_call_frame_cfa)
    00067d66 0000000000482fb8 00000000004831be (DW_OP_call_frame_cfa)
    00067d79 <End of list>

Which doesn't cover 0x482953 which is $GOROOT/src/fmt.print.go:682 where p.printValue is called.

Also this lexical block from the same function:

 <2><3149c>: Abbrev Number: 8 (DW_TAG_lexical_block)
    <3149d>   DW_AT_ranges      : 0x16280
 <3><314a1>: Abbrev Number: 5 (DW_TAG_variable)
    <314a2>   DW_AT_name        : f
    <314a4>   DW_AT_decl_line   : 677
    <314a6>   DW_AT_location    : 0x686a3 (location list)
    <314aa>   DW_AT_type        : <0x3403e>

the ranges for the lexical block are:
    00016280 ffffffffffffffff 00000000004827d0 (base address)
    00016280 000000000048292b 0000000000482941 
    00016280 0000000000482953 00000000004829ae 
    00016280 <End of list>

while the loclist for 'f' (this particular f, there are other fs in the function) is:

    000686a3 ffffffffffffffff 00000000004827d0 (base address)
    000686b3 00000000004827f1 0000000000482805 (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_piece: 8)
    000686cb 0000000000482805 000000000048290f (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    000686e5 00000000004829f7 0000000000482a35 (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    000686ff 0000000000482beb 0000000000482c29 (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    00068719 0000000000482cab 0000000000482cfa (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    00068733 0000000000482d64 0000000000482dd3 (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    0006874d 0000000000482e43 0000000000482e8b (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    00068767 0000000000482ef0 0000000000482f4d (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    00068781 0000000000482fb8 00000000004831be (DW_OP_fbreg: -16; DW_OP_piece: 8; DW_OP_fbreg: -16; DW_OP_piece: 8)
    0006879b <End of list>

once again PC 0x482953 isn't covered even though 'f' (this f) is evaluated on that line (it's the same as before)



Finally, and this is also not a regression, let's take this program https://play.golang.org/p/ux34T3xDJI and calculate the address of longintarr at 0x4861a9 (it's just before longintarr is passed to convT2Eslice).

The DIE of longintarr and main.main are:

 <1><3177a>: Abbrev Number: 2 (DW_TAG_subprogram)
    <3177b>   DW_AT_name        : main.main
    <31785>   DW_AT_low_pc      : 0x486120
    <3178d>   DW_AT_high_pc     : 0x4861f1
    <31795>   DW_AT_frame_base  : 1 byte block: 9c      (DW_OP_call_frame_cfa)
    <31797>   DW_AT_external    : 1
 <2><31798>: Abbrev Number: 5 (DW_TAG_variable)
    <31799>   DW_AT_name        : longintarr
    <317a4>   DW_AT_decl_line   : 4
    <317a5>   DW_AT_location    : 0x696c3 (location list)
    <317a9>   DW_AT_type        : <0x3834b>
 <2><317ad>: Abbrev Number: 0

the loclist is:

    000696c3 ffffffffffffffff 0000000000486120 (base address)
    000696d3 0000000000486151 0000000000486156 (DW_OP_reg5 (rdi))
    000696e6 0000000000486156 00000000004861f1 (DW_OP_fbreg: -16)

which for address 0x4861a9 gives us "DW_OP_fbreg -16"

The DFE for main.main is:

0001224c 0000000000000024 00000000 FDE cie=00000000 pc=0000000000486120..00000000004861f1
  DW_CFA_def_cfa_offset_sf: 8
  DW_CFA_advance_loc: 23 to 0000000000486137
  DW_CFA_def_cfa_offset_sf: 112
  DW_CFA_advance_loc1: 175 to 00000000004861e6
  DW_CFA_def_cfa_offset_sf: 8
  DW_CFA_advance_loc: 10 to 00000000004861f0

So for address 0x4861a9 the CFA offset is 112

RSP was for me 0xc420043f10 giving a call frame CFA of 0xc420043f80 which gives the address of longintarr as 0xc420043f70. I couldn't convince gdb to give me the calculated address of longintarr at that location but inspecting its contents looks like gdb agrees with me.

The problem is that 0xc420043f70 contains nonsense and at that PC address we have the real address of longintarr conveniently calculated in RAX: 0xc420043f58.

Patch-set: 4
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1