Update patch set 4
Patch Set 4:

> There are also a lot of cases for this with LastIndex(Byte). But that function is from go1.5, therefore I'm not sure what would be OK to change without breaking bootstrapping or other guarantees.

Technically I think you could make that replacement anywhere if you added "strings" to the list bootstrapDirs in cmd/dist/buildtool.go, but in practice that is an unnecessary hassle.  It's safe to use LastIndexByte outside of cmd/* and any packages imported by cmd/*.  For example, it would be safe to change the uses in the net/http package.

Patch-set: 4