cmd/compile: disallow loop preemption in the runtime
While we already disallow loop preemption for nosplit functions, there
are lots of runtime functions that "know" they won't be preempted
(like scheduler functions running on the scheduler stack and
signal-handling functions). It would be nice to have a go:nosched
annotation that was separate from go:nosplit (and probably deep), but
in lieu of that, just opt the whole runtime out of loop preemption.

Updates #10958.

Change-Id: I8cce75bffc08763308d9d8508be1b0e944f53554