runtime: keep rescheduling fault page together
There's no guarantee that the padding around the rescheduling fault
check byte will actually be colocated with the check byte. In fact,
it's not.

Since we really need this all together so we can unmap the whole page
containing the check byte, pack it all into a single struct instead of
using multiple global variables.

We also make the padding 64K since this is the largest physical page
size of any system we support (we'll add a runtime check to make sure
this is enough later) and move this into its own source file (which
will grow in the future).

Updates #10958.

Change-Id: I0d4cc03dbdd67cd33275310364ab0aa3d244f1a2