cmd/compile: record rescheduling target PC
This adds PCDATA at fault-based rescheduling test instruction to
inform the runtime of the PC of the rescheduling branch.

To make this possible, this also extends the PCDATA instruction to
accept a branch target, which gets recorded in the PCDATA tables as
the target PC relative to the entry point of the function.

Updates #10958.

Change-Id: I6f2a10ef6d2b472753c54f727c0955eadcc8e12d