cmd/compile, runtime: make bad safe-points throw
Currently, loop preemption is done by calling runtime.goschedguarded,
which simply returns if it's a bad safe-point time.

This is problematic for fault-based preemption since the caller will
just keep faulting on each subsequent iteration until it reaches a
good safe-point time.

The compiler shouldn't generate safe-points in any situation that this
might apply to, so make runtime.goschedguarded throw in these bad
situations. This also lifts the test to before the mcall so if it does
fail we get a real backtrace. Since the name is no longer really
appropriate, also rename it to runtime.safepoint.

Updates #10958.

Change-Id: I0c2cef22a5634b4a0769a3a394af3d1279c3f28b