cmd/trace: add /usertasks and /usertask views
This implements basic annotation analysis - grouping tasks, spans, and
goroutines, and building per-task type latency histogram.

/trace is also extended to provide task-oriented view of involved
goroutines when taskid param is specified. This is similar to the
goroutine-oriented trace view but instead of heuristically finding
related goroutines, it uses the grouping from the annotation analysis
and presents those goroutines along with task, and span information.

TODO:
  present subtasks together.
  analysis on task-less spans.

Change-Id: I0a0f6f32519b49d944d9386bc8be15ce45444f0d