Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (2 comments)
> 
> AFAIK gdb's next doesn't work too well with go, a preemption or a stack growth could make the test fail, I don't know if this is possible with this particular program.

This is meant to be built with optimizations on (I think). General goal is to make that work much better, many of the current mistakes are boneheaded.
This example seems to not have any stack growth in it, and I don't think (but I could be wrong) that delve (and in particular, a known version of delve) is installed on our test machines.

I'm unsure about whether I should delete the executable. I could add it to .gitignore.  I think the trybots each run in their own fresh container (at least, I hope that is how it works).

Patch-set: 7
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>