runtime: ignore loop preemption signals in GDB
Now that we're using SIGSEGV traps to preempt loops, GDB often stops
unexpectedly and mysteriously at loops. Teach GDB to ignore SIGSRGV
while we're preempting loops. This isn't perfect, since it could in
principle race with a real SIGSEGV, causing GDB to not trap on the
real SIGSEGV. It further isn't perfect because there could be pending
preemption SIGSEGVs when we re-enable trapping SIGSEGV. To handle the
latter, we use a second line of defense that detects preemption
SIGSEGVs and continues execution.

A much better solution would be to take the trap, examine the fault
address, and use this to silently resume from preemption signals.
Unfortunately, GDB prints a big message about the signal before
calling anything we control and there appears to be no way to suppress
this. Hence, we use this approach for our fallback, but it's not
practical as the only solution in GDB. For a native debugger, this
would obviously be the way to go.

Updates #10958.

Change-Id: Icc428a897651471f108c0ef7c1728c9b411038f2