Update patch set 2
Patch Set 2:

I'm not thrilled with dealing with duplicates on rename, but I think this is a good idea. It removes the build mode concept from the object file reader code, and that's a nice strand of spaghetti to straighten out. (This is clearer in a later CL that moves the object reader to its own package.)

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>