net/http/httptrace: expose transport-added request headers for http/1.1
Expose headers which are set or modified by the http library.
These headers are not written to the standard http.Request.Header
and are not included as part of http.Response.Request.Header.

Seems safe to pass the header map without copying because this is done at the end
of Request.write and therefore after the values have been written to the wire.
And it's not possible that these headers will be used on later requests because
a new transportRequest is created for each retry.

This is not a complete solution to 19761 since it does not have http/2 support.
http/2 appears to need support here solely due to its modification of Content-Length.

Fixes #19761

Change-Id: Ie8d4f702f4f671666b120b332378644f094e288b