Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Care to take a look at this? I'm not sure why we were testing for B.LS, but I've checked arm{5,6,7} and it seems we always generate a plain B for main.main
> 
> I am also confused with B.LS, which I thought should be a plain B. And there is indeed B in the disassembly.

Thanks. Since the disasm result seems to be correct, it doesn't make much sense to test for a 'B.LS' that we never generate, and it seems like we can fix the TODO by just changing the test to 'B' and un-commenting it.

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>