Update patch set 1
Patch Set 1:

Conceptually this is simple, but there are a number of edge cases to consider esp when dealing with adding and returning connection to the pool.

I'm working on adding tests and getting previous tests to pass when a connection resetter is present.

Patch-set: 1
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>