Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> LGTM once builders are happy.
> 
> Alex

Moved to before the CLs that introduce the build failures.

This can be submitted on its own, but it has to be in this stack somewhere since later changes depend on it. But I can't move it before David's CL at the head of the stack since that would rebase his CL. :)

Patch-set: 3
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>