cmd/link: suppress unnecessary DWARF relocs that confuse dsymutil
During Mach-O linking, dsymutil takes the DWARF from individual object
files and combines it into a debug archive. Because it's content-aware,
it doesn't need our help to do its job. Nonetheless, it does try to
honor relocations that are present in its input. It doesn't do it very
well.

In particular, dsymutil only understands relocation to functions, and it
doesn't actually check that it's processing such a relocation before
rewriting it. It just matches addresses. So if it happens to find a
ref_addr whose offset matches a function address, it will happily
rewrite that, producing nonsense.

Since the relocations don't help, and can hurt, drop them when
externally linking a Mach-O binary.

Fixes #22068

Change-Id: I8ec36da626575d9f6c8d0e7a0b76eab8ba22d62c