cmd/link: suppress unnecessary DWARF relocs that confuse dsymutil
During Mach-O linking, dsymutil takes the DWARF from individual object
files and combines it into a debug archive. Because it's content-aware,
it doesn't need our help to do its job. Nonetheless, it does try to
honor relocations that are present in its input.

dsymutil reads the addends of the relocations present in the objects
it's given, and uses those as an index into the debug map to find their
final location. When it does that, it's assuming that the value is an
address in the object file. But DWARF references are section-relative.
So when it processes a relocation for a DWARF reference, it gets
confused, and if the value happens to match the address of a function or
data symbol, it will rewrite it incorrectly.

Since the relocations don't help, and can hurt, drop them when
externally linking a Mach-O binary.

Fixes #22068

Change-Id: I8ec36da626575d9f6c8d0e7a0b76eab8ba22d62c