Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> The LoadObjFile variable is pretty ugly.  Is this really a good way to split up the package?

Instead of attempting to describe the plan I sketched out, I thought it would be easier to demonstrate part of it. So I rebased this on top of CL 68331. Please let me know what you think of that.

I don't know what the ideal number of packages for the final form of the linker is, but I believe as intermediate steps it is valuable to separate out parts of the linker into a package, to help unravel the wound together code. For example, now we know the object reading code doesn't depend on any flag values.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>