Update patch set 4
Patch Set 4:

I'm very sad to see this. I have been consistently encouraging people to use strings.Index(s, "x") instead of strings.IndexByte(s, 'x') because the former is clearer, more general, and just as fast. 

This CL is making the code more complex instead of more readable, and for no apparent benefit - there was no accompanying compelling measurement or anything like that.

I would like to revert this change. This is not how we should be writing Go code.

Patch-set: 4
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>