Update patch set 4
Patch Set 4:

With all do respect, if I may, it seems a bit dramatic. Using IndexByte seems more specific to that specific call, even if the presence of It may not be favoured. For now, it is part of the API. If there are plans to remove it then it would make sense to not use it. Just my two cents because I was subscribed to this thread.

On Thu, 5 Oct 2017 at 16:35, Marvin Stenger (Gerrit) <> wrote:

Patch-set: 4
Tag: mailMessageId=<CALh=o-xGd5js-YcLaK5zrq+fVkaZgmN7XdXu6wKZp-U3fziydg@mail.gmail.com>