Update patch set 1
Patch Set 1:

Sorry everyone, this should have said DO NOT REVIEW like all of the others below it in the stack. It's not ready yet and won't be for some time; it's only here to give a sense of where things are hopefully headed. I'm still trying to make everything leading up to here work for 100% of existing use cases, and then I'll revise this.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>