Update patch set 2
Patch Set 2:

> A debugger on macOS can do the right thing and allow you to resume from SIGSEGV. At the time I filed that bug, lldb did not do the right thing. See comment #1 from Jim Ingham. It probably still doesn't.

I can confirm that neither lldb, nor debugserver (which delve is using as a backend on macOS) is still not doing the right thing.

We have a native backend in delve but we have abandoned it because (a) we couldn't figure out how to use the undocumented debugging APIs correctly (b) not having to explain to users how to sign the executable is great for usability.

Patch-set: 2