cmd/go: stop treating GOROOT binaries and runtime/cgo specially
This was for cmd/cgo, and cmd/cgo isn't special anymore.

Now that we don't treat GOROOT binaries specially, we
don't have to try to guess whether the linker wants external
linking - we can just provide runtime/cgo to all links and
leave it up to the linker to decide.

And we can delete all the code that was special-casing
GOROOT for external linking decisions.

Change-Id: I71efaf553b24798b523c7102859428d36b470698