Update patch set 1
Patch Set 1: Run-TryBot+1 Code-Review+2

(1 comment)

Is this the only type for which this is a problem? If not, we could consider adding a type for gathering up the types the heap dump reader needs, rather than scattering them. But this is the only type where this is a problem, then this is fine.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2
Label: Run-TryBot=+1