runtime: rename sched.mcount to sched.mnext
Currently, since Ms never exit, the number of Ms, the number of Ms
ever created, and the ID of the next M are all the same. That's about
to change, so rename sched.mcount to sched.mnext to make it clear it's
the number of Ms ever created (and the ID of the next M), and use
mcount() for the number of Ms. In the next commit, mcount() will
become slightly less trivial.

For #20395.

Change-Id: I9af34d36bd72416b5656555d16e8085076f1b196