runtime: fix cgo export of panicmem on ios
CL 68490 reworked the way the panicmem function is exposed to the
C mach expection catcher. However //go:cgo_export_static isn't enough:
the function name has to be equal to the remote symbol, and even so,
//go:linkname is needed. Finally, exporting an assembler function
doesn't work, so a Go wrapper is added.

Also, mark the overridden C symbol weak to to avoid duplicate symbol
errors from the host linker.

For the ios builders.

Change-Id: Ib87789fecec9314e398cf1bd8c04ba0b3a6642af