cmd/vet: skip self-assigns that contain any calls
The existing logic for whether the left and right parts of an assignment
were equal only checked that the gofmt representation of the two was
equal. This only checks that the ASTs were equal.

However, that method is flawed. For example, if either of the
expressions contains a function call, the expressions may actually be
different even if their ASTs are the same. An obvious case is a func
call to math/rand to get a random integer, such as the one added in the
test.

If either of the expressions contains a func call, simply skip the
check.

Fixes #22174.

Change-Id: Ied7f7543dc2bb8852e817230756c6d23bc801d90