Update patch set 1
Patch Set 1:

This is blocked by the fact that no trybot on builder has upx installed (so the test would be skipped everywhere). If we are interested in testing upx compression (I assume so: there a go1.10 open issue for that, see the commit description), we should equip a few trybots and builders with upx.

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>
Work-in-progress: false