Update patch set 1
Patch Set 1:

I'm not entirely sure if this is ok (but hopefully reviewers are). The documentation for the compile command (https://golang.org/cmd/compile/#hdr-Compiler_Directives) says that this directive affects "[...] the next declaration in the file [...]", which is making me think that it doesn't have to be adjacent to the declaration.

Running the tests locally on OS X (./all.bash) pass without any problems even when removing this directive. So, the tests passing isn't a confirmation that this hasn't broken anything.

The TryBots did fail because of a timeout, but I'm not sure if that's an indication of anything being wrong with the CL or if it's just something that happens once in a while.

Patch-set: 1
Reviewer: Gerrit User 5465 <5465@62eb7196-b449-3ce5-99f1-c037f21e1705>