cmd/go: skip "no buildable go files" in fmt
Otherwise, one can't run "go fmt" on a directory containing Go files if
none of them are buildable (e.g. because of build tags). This is
counter-intuitive, as fmt will format all Go files anyway.

If we encounter such a load error, ignore it and carry on. To detect the
nature of the underlying error, PackageError.Err's type was changed from
string to error.

Fixes #22183.

Change-Id: I9bebc6a82b0658f52776aa2928938b22d24bccdd