Update patch set 2
Patch Set 2:

> Ian (or Russ), in light of Russ' comments, should I still add the weak attribute to gcc_signal_darwin_armx.c?

The point of the weak attribute was to make it so that the cgo link did not care that xx_cgo_panicmem is not defined anywhere.  I don't know why it didn't work.  If you are keeping the definition in gcc_signal2_darwin_armx.c, then you don't need the weak attribute.

Patch-set: 2