Update patch set 2
Patch Set 2:

Thanks for the CL.

Can we leave this to the side for the moment? I have a bunch of pending changes in the go command and can return to thinking about how this should happen once those changes have settled some more. It may be that we should just have a flag to load that disables this check, for example, but I'm not sure at the moment, and there are lots of other things up in the air.

Please ping this in a couple weeks if you haven't heard from me, but I hope that won't be the case. Thanks.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>